Rename ftCo_ItemGet to ftpickupitem based on assert string
