+++ asm/melee/ft/chara/ftCommon/ftpickupitem.s
