+++ config/GALE01/symbols.txt
@@ -2312,10 +2312,10 @@ ftCo_80094694 = .text:0x80094694; // type:function size:0xFC scope:global
 ftCo_80094790 = .text:0x80094790; // type:function size:0x88 scope:global
 ftCo_80094818 = .text:0x80094818; // type:function size:0x90 scope:global
 ftCo_800948A8 = .text:0x800948A8; // type:function size:0x16C scope:global
-ftCo_ItemGet_Anim = .text:0x80094A14; // type:function size:0x10C scope:global
-ftCo_ItemGet_IASA = .text:0x80094B20; // type:function size:0x4 scope:global
-ftCo_ItemGet_Phys = .text:0x80094B24; // type:function size:0x20 scope:global
-ftCo_ItemGet_Coll = .text:0x80094B44; // type:function size:0x28 scope:global
+ftpickupitem_Anim = .text:0x80094A14; // type:function size:0x10C scope:global
+ftpickupitem_IASA = .text:0x80094B20; // type:function size:0x4 scope:global
+ftpickupitem_Phys = .text:0x80094B24; // type:function size:0x20 scope:global
+ftpickupitem_Coll = .text:0x80094B44; // type:function size:0x28 scope:global
 ftCo_80094B6C = .text:0x80094B6C; // type:function size:0x224 scope:local
 ftCo_80094D90 = .text:0x80094D90; // type:function size:0x68 scope:global
 ftCo_80094DF8 = .text:0x80094DF8; // type:function size:0x5C scope:global
