+++ configure.py
@@ -474,7 +474,7 @@ config.libs = [
             Object(NonMatching, "melee/ft/chara/ftCommon/ftCo_DamageFall.c"),
             Object(NonMatching, "melee/ft/chara/ftCommon/ftCo_DamageIce.c"),
             Object(NonMatching, "melee/ft/chara/ftCommon/ftCo_Guard.c"),
-            Object(NonMatching, "melee/ft/chara/ftCommon/ftCo_ItemGet.c"),
+            Object(NonMatching, "melee/ft/chara/ftCommon/ftpickupitem.c"),
             Object(NonMatching, "melee/ft/chara/ftCommon/ftCo_ItemThrow.c"),
             Object(Matching, "melee/ft/chara/ftCommon/ftCo_SpecialS.c"),
             Object(Matching, "melee/ft/chara/ftCommon/ftCo_SpecialAir.c"),
