+++ src/melee/ft/chara/ftCommon/ftCo_Attack1.c
@@ -8,9 +8,9 @@
 #include "ftCo_AttackLw4.h"
 #include "ftCo_AttackS3.h"
 #include "ftCo_AttackS4.h"
-#include "ftCo_ItemGet.h"
 #include "ftCo_ItemThrow.h"
 #include "ftCo_Wait.h"
+#include "ftpickupitem.h"
 
 #include "ft/fighter.h"
 #include "ft/ft_081B.h"
