+++ src/melee/ft/chara/ftCommon/ftCo_AttackLw3.c
@@ -11,8 +11,8 @@
 #include "ftCo_AttackLw4.h"
 #include "ftCo_AttackS3.h"
 #include "ftCo_AttackS4.h"
-#include "ftCo_ItemGet.h"
 #include "ftCo_ItemThrow.h"
+#include "ftpickupitem.h"
 
 #include "ft/fighter.h"
 #include "ft/ft_081B.h"
