+++ src/melee/ft/chara/ftCommon/ftCo_AttackS3.c
@@ -4,9 +4,9 @@
 
 #include "ftCo_AttackS3.h"
 
-#include "ftCo_ItemGet.h"
 #include "ftCo_ItemThrow.h"
 #include "ftCo_Wait.h"
+#include "ftpickupitem.h"
 #include "math.h"
 
 #include "ft/fighter.h"
