+++ src/melee/ft/chara/ftFox/ftFx_SpecialN.c
@@ -17,7 +17,7 @@
 #include "ft/ftparts.h"
 #include "ft/types.h"
 #include "ftCommon/ftCo_FallSpecial.h"
-#include "ftCommon/ftCo_ItemGet.h"
+#include "ftCommon/ftpickupitem.h"
 #include "ftFox/types.h"
 #include "it/it_26B1.h"
 #include "it/items/it_27CF.h"
