+++ src/melee/ft/chara/ftLink/ftLk_SpecialLw.c
@@ -14,8 +14,8 @@
 #include "ft/ftcommon.h"
 #include "ft/ftparts.h"
 #include "ft/types.h"
-#include "ftCommon/ftCo_ItemGet.h"
 #include "ftCommon/ftCo_ItemThrow.h"
+#include "ftCommon/ftpickupitem.h"
 #include "it/it_26B1.h"
 #include "it/items/it_27CF.h"
 #include "lb/lb_00B0.h"
