+++ src/melee/ft/chara/ftPeach/ftPe_SpecialLw.c
@@ -16,8 +16,8 @@
 #include "ft/ftanim.h"
 #include "ft/ftcommon.h"
 #include "ft/types.h"
-#include "ftCommon/ftCo_ItemGet.h"
 #include "ftCommon/ftCo_ItemThrow.h"
+#include "ftCommon/ftpickupitem.h"
 #include "ftPeach/types.h"
 #include "it/it_26B1.h"
 #include "it/items/it_27CF.h"
